refactor: move agent api key service into monitoring feature
